src/requirements-etl.txt
python-evtx
xmltodict
pandas
sqlalchemy
psycopg2-binary
python-dotenv
